@@ -3,7 +3,10 @@ from __future__ import annotations
 
 import argparse
 import os
-from typing import Sequence
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from collections.abc import Sequence
 
 
 class Tree:
